added build files and misc scripts
Added some stuff to build the project along with a start script to
start up the server (running as an akka kernel package). Also added
a custom sbt script to the project for those that may not have it
installed on their machine, and so I can customize it if necessary.